incremento para detalhes de alunos
